hw/mem/cxl_type3: improve qmp interfaces to add/release dc extents
Before the change, we are only able to pass number of extents to
operate and start dpa and extent length are hard coded.
With the commit, we improve the interface and allowing users to control
start dpa, extent length, region id other than number of extents to
operate.
The new interface looks like:

1. Add 2 extents with starting dpa 0 and 128MB extent length
{ "execute": "qmp_capabilities" }

{ "execute": "cxl-add-dynamic-capacity-event",
	"arguments": {
		 "path": "/machine/peripheral/cxl-pmem0",
		"region-id" : 0,
		 "num-extent": 2,
		"dpa":0,
		"extent-len": 128
	}
}

2. release the extent starting at dpa 0 and length 128MB

{ "execute": "cxl-release-dynamic-capacity-event",
	"arguments": {
		 "path": "/machine/peripheral/cxl-pmem0",
		"region-id" : 0,
		 "num-extent": 1 ,
		"dpa":0,
		"extent-len": 128
	}
}