hw/mem/cxl_type3: add bitmap to track whether a dc block is backed by an online extent
When a dc extent is added, set the bits of the bitmap where the blocks are
covered; when a dc extent is released, clear the bits of the bitmap
where the blocks are covered.
During dc read/write, check whether the dpa range is backed by onlined
extents by checking the bitmap.