hw/mem/cxl_type3: Add a parameter to pass number of DC regions the device supports in qemu command line
Add a property 'num-dc-regions' to ct3_props to allow users to create DC
regions.
With the change, users can control the number of DC regions the device
supports.
To make it easier, other parameters of the region like region base, length,
and block size are hard coded. If desired, these parameters
can be added easily.

Signed-off-by: Fan Ni <fan.ni@samsung.com>